+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `step_texthash()` and `step_dummy_hash()` gained `sparse` argument. When set to `"yes"`, `step_dummy()` will produce sparse vectors. (#282)
 
+* `step_tf()` gained `sparse` argument. When set to `"yes"`, `step_dummy()` will produce sparse vectors. (#284)
+
 # textrecipes 1.0.7
 
 ## Improvements